+++ views/edit.ejs
@@ -1,0 +1,16 @@
+<!DOCTYPE html>
+<html>
+   <head>
+      <title>Cadastro de planos</title>
+   </head>
+   <body>
+      <h2>Cadastro de planos</h2>
+      <form action="/planos" method="post">
+         <label for="fname">Nome do plano:</label><br>
+         <input type="text" id="name" name="name" placeholder="Nome do plano" value="<%= planos.nome %>"><br>
+         <label for="lname">Preço:</label><br>
+         <input type="number" id="preco" name="preco" placeholder="Preço do plano" value="<%= planos.preco %>" ><br><br>
+         <input type="submit" value="atualizar">
+      </form>
+   </body>
+</html>
