+++ views/index.ejs
@@ -35,12 +35,15 @@
          <% Planos.forEach(plano => {%>
          <tr>
             <td> <%= plano.id %></td>
-            <td> <%= plano.name %></td>
+            <td> <%= plano.nome %></td>
             <td> <%= plano.preco %></td>
             <td id="td-acoes">
                <div id="acoes">
-                  <button>Editar</button>
-                  <button>Excluir</button>
+                 <a href="/planos/<%= plano.id %>"> <button>Editar</button></a>
+                 <form method="POST" action="/planos/delete" style="display: inline;" onsubmit="confirmarDelecao(event,this)">
+                  <input type="hidden" name="id" value="<%= plano.id %>">
+                  <button class="button"">Delete</button>
+              </form>
                </div>
             </td>
          </tr>
